source code cleanup
git-svn-id: https://findbugs.googlecode.com/svn/trunk@12688 eae3c2d3-9b19-0410-a86e-396b6ccb6ab3

+++ findbugsTestCases/src/java/bugIdeas/Ideas_2008_08_06.java
@@ -18,7 +18,7 @@ public class Ideas_2008_08_06<E> extends HashSet<E> {
     @ExpectWarning("GC,Dm")
     public static void main(String args[]) {
         Ideas_2008_08_06<String> i = new Ideas_2008_08_06<String>();
-        System.out.println(i.contains((Integer) 5));
+        System.out.println(i.contains(5));
         ScheduledThreadPoolExecutor e = new ScheduledThreadPoolExecutor(0);
         e.setMaximumPoolSize(10);
     }
