+++ findbugsTestCases/src/java/bugPatterns/ICAST_IDIV_CAST_TO_DOUBLE.java
@@ -3,7 +3,7 @@ package bugPatterns;
 public class ICAST_IDIV_CAST_TO_DOUBLE {
 
     void bug(int x, int y) {
-        double d = (double) (x / y);
+        double d = (x / y);
     }
 
 }
